+++ include/Initializers/XavierInitializer.h
@@ -11,14 +11,14 @@ public:
     using ParallelInitializer::ParallelInitializer;
 
 protected:
-    void fill_chunk(std::span<fp16> chunk, uint64_t input_dim, uint64_t output_dim, uint32_t chunk_seed) const override {
+    void fill_chunk(std::span<fp32> chunk, uint64_t input_dim, uint64_t output_dim, uint32_t chunk_seed) const override {
         fp32 sigma = std::sqrt(1.0f / static_cast<fp32>(input_dim));
         
         std::mt19937 local_gen(chunk_seed);
         std::normal_distribution<fp32> local_dist(0.0f, sigma);
 
         for (auto& w : chunk) {
-            w = static_cast<fp16>(local_dist(local_gen));
+            w = static_cast<fp32>(local_dist(local_gen));
         }
     }
 };
